+++ MapboxCoreNavigation/LegacyRouteController.swift
@@ -87,7 +87,12 @@ open class LegacyRouteController: NSObject, Router, InternalRouter, CLLocationMa
 
     var userSnapToStepDistanceFromManeuver: CLLocationDistance?
     
-    required public init(along route: Route, routeIndex: Int, options: RouteOptions, directions: Directions = Directions.shared, dataSource source: RouterDataSource) {
+    required public init(along route: Route,
+                         routeIndex: Int,
+                         options: RouteOptions,
+                         directions: Directions = Directions.shared,
+                         dataSource source: RouterDataSource,
+                         tilesVersion: String? = nil) {
         self.directions = directions
         self._routeProgress = RouteProgress(route: route, routeIndex: routeIndex, options: options)
         self.dataSource = source