+++ MapboxCoreNavigation/NavigationService.swift
@@ -215,7 +215,11 @@ public class MapboxNavigationService: NSObject, NavigationService {
      - parameter routeindex: The index of the route within the original `RouteController` object.
      */
     convenience init(route: Route, routeIndex: Int, routeOptions options: RouteOptions) {
-        self.init(route: route, routeIndex: routeIndex, routeOptions: options, directions: nil, locationSource: nil, eventsManagerType: nil)
+        self.init(route: route,
+                  routeIndex: routeIndex,
+                  routeOptions: options, directions: nil,
+                  locationSource: nil,
+                  eventsManagerType: nil)
     }
     
     /**
@@ -228,6 +232,7 @@ public class MapboxNavigationService: NSObject, NavigationService {
      - parameter eventsManagerType: An optional events manager type to use while tracking the route.
      - parameter simulationMode: The simulation mode desired.
      - parameter routerType: An optional router type to use for traversing the route.
+     - parameter tilesVersion: Version of tiles downloaded via Offline Service.
      */
     required public init(route: Route,
                          routeIndex: Int,
@@ -236,7 +241,8 @@ public class MapboxNavigationService: NSObject, NavigationService {
                          locationSource: NavigationLocationManager? = nil,
                          eventsManagerType: NavigationEventsManager.Type? = nil,
                          simulating simulationMode: SimulationMode = .onPoorGPS,
-                         routerType: Router.Type? = nil) {
+                         routerType: Router.Type? = nil,
+                         tilesVersion: String? = nil) {
         nativeLocationSource = locationSource ?? NavigationLocationManager()
         self.directions = directions ?? Directions.shared
         self.simulationMode = simulationMode
@@ -249,7 +255,12 @@ public class MapboxNavigationService: NSObject, NavigationService {
         }
         
         let routerType = routerType ?? DefaultRouter.self
-        router = routerType.init(along: route, routeIndex: routeIndex, options: routeOptions, directions: self.directions, dataSource: self)
+        router = routerType.init(along: route,
+                                 routeIndex: routeIndex,
+                                 options: routeOptions,
+                                 directions: self.directions,
+                                 dataSource: self,
+                                 tilesVersion: tilesVersion)
         NavigationSettings.shared.distanceUnit = routeOptions.locale.usesMetric ? .kilometer : .mile
         
         let eventType = eventsManagerType ?? NavigationEventsManager.self