+++ MapboxCoreNavigation/OfflineDirections.swift
@@ -11,7 +11,6 @@ public typealias NavigationDirectionsCompletionHandler = (_ tilesURL: URL) -> Vo
 /** :nodoc:
  An error that occurs when calculating directions potentially offline using the `NavigationDirections.calculate(_:offline:completionHandler:)` method.
 */
-@available(*, deprecated)
 public enum OfflineRoutingError: LocalizedError {
     /**
      A standard Directions API error occurred.
@@ -60,7 +59,6 @@ public enum OfflineRoutingError: LocalizedError {
     }
 }
 
-@available(*, deprecated)
 struct NavigationDirectionsConstants {
     static let offlineSerialQueueLabel = Bundle.mapboxCoreNavigation.bundleIdentifier!.appending(".offline")
     static let unpackSerialQueueLabel = Bundle.mapboxCoreNavigation.bundleIdentifier!.appending(".offline.unpack")
@@ -97,7 +95,6 @@ public typealias UnpackCompletionHandler = (_ numberOfTiles: UInt64, _ error: Er
  If the request was canceled or there was an error obtaining the routes, this argument is `nil`. This is not to be confused with the situation in which no results were found, in which case the array is present but empty.
  - parameter error: The error that occurred, or `nil` if the placemarks were obtained successfully.
  */
-@available(*, deprecated)
 public typealias OfflineRouteCompletionHandler = (_ session: Directions.Session, _ result: Result<RouteResponse, OfflineRoutingError>) -> Void
 
 /** :nodoc:
@@ -105,14 +102,14 @@ public typealias OfflineRouteCompletionHandler = (_ session: Directions.Session,
  
  Each result produced by the directions object is stored in a `Route` object. Depending on the `RouteOptions` object you provide, each route may include detailed information suitable for turn-by-turn directions, or it may include only high-level information such as the distance, estimated travel time, and name of each leg of the trip. The waypoints that form the request may be conflated with nearby locations, as appropriate; the resulting waypoints are provided to the closure.
  */
-@available(*, deprecated, message: "Use the Directions class instead.")
 public class NavigationDirections: Directions {
     /**
      Configures the router with the given set of tiles.
      
-     - parameter tilesURL: The location where the tiles has been sideloaded to.
+     - parameter tilesURL: The location where the tiles have been sideloaded to.
      - parameter completionHandler: A block that is called when the router is completely configured.
      */
+    @available(*, deprecated)
     public func configureRouter(tilesURL: URL, completionHandler: @escaping NavigationDirectionsCompletionHandler) {
         NavigationDirectionsConstants.offlineSerialQueue.sync {
             let tilesConfig = TilesConfig(tilesPath: tilesURL.path,
@@ -131,6 +128,25 @@ public class NavigationDirections: Directions {
     }
     
     /**
+     Configures the router with the given `tilesVersion`.
+     
+     - parameter tilesVersion: Version of tiles downloaded via Offline Service.
+     */
+    public func configureRouter(tilesVersion: String) {
+        NavigationDirectionsConstants.offlineSerialQueue.sync {
+            let endpointConfig = TileEndpointConfiguration(directions: Directions.shared, tilesVersion: tilesVersion)
+            let tilesConfig = TilesConfig(tilesPath: Bundle.mapboxCoreNavigation.suggestedTileURL(version: tilesVersion)?.path ?? "",
+                                          inMemoryTileCache: nil,
+                                          mapMatchingSpatialCache: nil,
+                                          threadsCount: nil,
+                                          endpointConfig: endpointConfig)
+            
+            let settingsProfile = SettingsProfile(application: ProfileApplication.kMobile, platform: ProfilePlatform.KIOS)
+            self.navigator = Navigator(profile: settingsProfile, config: NavigatorConfig() , customConfig: "", tilesConfig: tilesConfig)
+        }
+    }
+    
+    /**
      Unpacks a .tar-file at the given filePathURL to a writeable output directory.
      The target at the filePathURL will be consumed while unpacking.
      
@@ -139,6 +155,7 @@ public class NavigationDirections: Directions {
      - parameter progressHandler: Unpacking reports progress every 500ms.
      - parameter completionHandler: Called when unpacking completed.
      */
+    @available(*, deprecated)
     public class func unpackTilePack(at filePathURL: URL, outputDirectoryURL: URL, progressHandler: UnpackProgressHandler?, completionHandler: UnpackCompletionHandler?) {
         NavigationDirectionsConstants.offlineSerialQueue.sync {
             let totalPackedBytes = filePathURL.fileSize!