+++ MapboxNavigation/NavigationViewController.swift
@@ -255,7 +255,10 @@ open class NavigationViewController: UIViewController, NavigationStatusPresenter
     required public init(for route: Route, routeIndex: Int, routeOptions: RouteOptions, navigationOptions: NavigationOptions? = nil) {
         super.init(nibName: nil, bundle: nil)
         
-        self.navigationService = navigationOptions?.navigationService ?? MapboxNavigationService(route: route, routeIndex: routeIndex, routeOptions: routeOptions)
+        self.navigationService = navigationOptions?.navigationService ?? MapboxNavigationService(route: route,
+                                                                                                 routeIndex: routeIndex,
+                                                                                                 routeOptions: routeOptions,
+                                                                                                 tilesVersion: navigationOptions?.tilesVersion)
         self.navigationService.delegate = self
 
         let credentials = navigationService.directions.credentials